use uint112 for invested amounts

+++ test/__mocks/AssetManager.sol
@@ -8,30 +8,30 @@ import { IAssetManager } from "src/interfaces/IAssetManager.sol";
 
 contract AssetManager is IAssetManager
 {
-    mapping(address => mapping(address => uint256)) public getBalance;
+    mapping(address => mapping(address => uint112)) public getBalance;
 
-    function adjustInvestment(IUniswapV2Pair aPair, int256 aToken0Amount, int256 aToken1Amount) external
+    function adjustInvestment(IUniswapV2Pair aPair, int224 aToken0Amount, int224 aToken1Amount) external
     {
-        require(aToken0Amount != type(int256).min && aToken1Amount != type(int256).min, "overflow");
+        require(aToken0Amount != type(int224).min && aToken1Amount != type(int224).min, "overflow");
 
         if (aToken0Amount >= 0) {
-            uint256 lAbs = uint256(aToken0Amount);
+            uint112 lAbs = uint112(uint256(int256(aToken0Amount)));
 
             getBalance[address(aPair)][aPair.token0()] += lAbs;
         }
         else {
-            uint256 lAbs = uint256(-aToken0Amount);
+            uint112 lAbs = uint112(uint256(int256(-aToken0Amount)));
 
             IERC20(aPair.token0()).approve(address(aPair), lAbs);
             getBalance[address(aPair)][aPair.token0()] -= lAbs;
         }
         if (aToken1Amount >= 0) {
-            uint256 lAbs = uint256(aToken1Amount);
+            uint112 lAbs = uint112(uint256(int256(aToken1Amount)));
 
             getBalance[address(aPair)][aPair.token1()] += lAbs;
         }
         else {
-            uint256 lAbs = uint256(-aToken1Amount);
+            uint112 lAbs = uint112(uint256(int256(-aToken1Amount)));
 
             IERC20(aPair.token1()).approve(address(aPair), lAbs);
             getBalance[address(aPair)][aPair.token1()] -= lAbs;
@@ -40,7 +40,7 @@ contract AssetManager is IAssetManager
         aPair.adjustInvestment(aToken0Amount, aToken1Amount);
     }
 
-    function adjustBalance(address aOwner, address aToken, uint256 aNewAmount) external
+    function adjustBalance(address aOwner, address aToken, uint112 aNewAmount) external
     {
         getBalance[aOwner][aToken] = aNewAmount;
     }
